pyproject.toml
[tool.poetry]
name = "discord-linking"
version = "0.1.0"
description = ""
authors = ["Alex Krantz <alex@krantz.dev>"]
readme = "README.md"
packages = [{include = "discord_linking"}]

[tool.poetry.dependencies]
python = "^3.9"
Flask = { version = "^2.1.2", extras = ["async"] }
python-dotenv = "^0.20.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
psycopg2 = { version = "^2.9.3", markers = "sys_platform == 'linux'" }
psycopg2-binary = { version = "^2.9.3", markers = "sys_platform != 'linux'" }
Authlib = "^1.0.1"
requests = "^2.27.1"
aws-requests-auth = "^0.4.3"
botocore = "^1.26.5"
Flask-Caching = "^1.10.1"
redis = "^4.3.1"
gunicorn = "^20.1.0"
opentelemetry-api = "^1.11.1"
opentelemetry-sdk = "^1.11.1"
opentelemetry-exporter-jaeger = "^1.11.1"
opentelemetry-exporter-otlp-proto-grpc = "^1.11.1"
opentelemetry-instrumentation-flask = "0.30b1"
opentelemetry-instrumentation-requests = "0.30b1"
opentelemetry-instrumentation-sqlalchemy = "0.30b1"
opentelemetry-instrumentation-botocore = "0.30b1"
opentelemetry-instrumentation-jinja2 = "0.30b1"
opentelemetry-instrumentation-redis = "0.30b1"
nats-py = "^2.1.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
ipython = "^8.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
